BUG FIX Needs Testing: When a call is dialling outside the system (using the dialler or someone's phone) the panel gets set as the user who dialled.  What we really want is to change this to the details of the person we're calling.
This bug meant that when the call was answered, if you filled out all the information, the details were lost as soon as they were put on hold (as the panel became the receiver and not the dialler).  Simple Bug fix by swapping the numbers around.

This creates more bugs: Panel no longer responds to conversation and other updates.  Suspect I need to do a channel flip on the internal data...