Tracks the onairQueue number so that we can see if a call may need to have its channel swapped (as it was a call we dialled out) in the CallLog
BUG FIX: Its possible to receive an updatefield command for a row that
doesn't exist yet.  So we need to create the row if it doesn't exist.

BUG FIX: Swap channel on dialled calls.  If QUEUE is on air queue then
swap the channel of an existing log entry to its new channel (Same BUG
as the CallManagerPanel due to how Asterisk works)

Added: void swapChannel(String, String): Swaps the channel for a given
log (denoted by the original channel) and updates the log table
accordingly