Removal of outgoing work around code, major change needs full testing
Removed all tbe outgoing call tracking as it didn't make sense anymore.
We should be able to track as all CALLs from an internal number to 
somewhere else will behave the same way.

We need to implement a channel swap from the dialler to the receiver 
and then track this call if it enters the queue as a new permanent
channel will be created for the rest of the life of the call.

Removed command[1] == UNKNOWN, we've cut this off at the server end and
obviously no need to process on the client.