Outgoing call bugfixes needs all call testing!
Implemented a whole bunch of work arounds for outgoing calls.  This 
mainly revolves around substituting outbound trunk caller ids on the
CONNECTED messages so that external outbound behaves the same as
internal outbound. 

I have no idea if this functions the same everywhere or is peculiar to 
our production system.

Suppressed double sided CONNECTED on outgoing calls, this will leave
clients in a RINGING state for outgoing calls.  So need to check and
remove this as appropriate upon receiving a CONNECTED message. TODO

Also implemented calls and ringingExternal HashMap cleanups on HANGUP