Moved Outside Call coming in to a queue into void createIncomingCall
Added call creation to queue handling if the panel does not exist.  This
helps to mitigate the problem when some trunks report incoming calls to
the trunk CID rather than the end point CID (e.g. the incoming queue).

This may not work for an external call direct to an internal phone
however it is not something we need to use with this system at the
moment.