Update AsteriskManager.java
Smack 4 updates.

PacketListener -> MessageListener

Added toString() to convert from ResourcePart.  This needs testing.

Removed old MessageListener as this is deprecated and no longer required as we don't listen for incoming private messages.

This was previously required due to how createPrivateChat worked.  We can skip all of this with ChatManager.chatWith().